Migrate modules and utils from AOS PYEZ to RESTapi
pylint cleanup
