+++ library/old_modules/aos_blueprint.py
@@ -133,7 +133,7 @@ build_errors:
 '''
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.aos import get_aos_session, check_aos_version, find_collection_item
+from ansible.module_utils.aos import get_aos_session, find_collection_item
 from ansible.module_utils.pycompat24 import get_exception
 
 def create_blueprint(module, aos, name):
@@ -286,9 +286,6 @@ def main():
         required_one_of=[('name', 'id')],
         supports_check_mode=True
     )
-
-    # Check if aos-pyez is present and match the minimum version
-    check_aos_version(module, '0.6.0')
 
     aos_blueprint(module)
