+++ library/old_modules/aos_blueprint_param.py
@@ -162,7 +162,7 @@ params_list:
 import json
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.aos import get_aos_session, find_collection_item, check_aos_version
+from ansible.module_utils.aos import get_aos_session, find_collection_item
 from ansible.module_utils.pycompat24 import get_exception
 
 try:
@@ -363,9 +363,6 @@ def main():
         ),
         supports_check_mode=True
     )
-
-    # Check if aos-pyez is present and match the minimum version
-    check_aos_version(module, '0.6.0')
 
     # aos-pyez availability has been verify already by "check_aos_version"
     # but this module requires few more object
