+++ library/old_modules/aos_blueprint_virtnet.py
@@ -86,7 +86,7 @@ import json
 
 from ansible.module_utils.pycompat24 import get_exception
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.aos import get_aos_session, find_collection_item, do_load_resource, check_aos_version, content_to_dict
+from ansible.module_utils.aos import get_aos_session, find_collection_item, do_load_resource, content_to_dict
 
 def ensure_present(module, aos, blueprint, virtnet):
 
@@ -206,9 +206,6 @@ def main():
         required_one_of=[('name', 'content')],
         supports_check_mode=True
     )
-
-    # Check if aos-pyez is present and match the minimum version
-    check_aos_version(module, '0.6.0')
 
     blueprint_virtnet(module)
