+++ library/old_modules/aos_hcl.py
@@ -18,24 +18,23 @@
 # along with Ansible.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-ANSIBLE_METADATA = {'metadata_version': '1.0',
-                    'status': ['preview'],
-                    'supported_by': 'community'}
-
+ANSIBLE_METADATA = {'status': ['preview'],
+                    'supported_by': 'community',
+                    'version': '1.0'}
 
 DOCUMENTATION = '''
 ---
-module: aos_rack_type
-author: Damien Garros (@dgarros)
-version_added: "2.3"
-short_description: Manage AOS Rack Type
+module: aos_hcl
+author: Jeremy Schulman (@jeremyschulman)
+version_added: "2.4"
+short_description: Manage AOS Hardware Compatibility List
 description:
-  - Apstra AOS Rack Type module let you manage your Rack Type easily.
-    You can create create and delete Rack Type by Name, ID or by using a JSON File.
+  - Apstra AOS HCL module let you manage your Hardware entries easily.
+    You can create create and delete an item by Name, ID or by using a JSON File.
     This module is idempotent and support the I(check) mode.
     It's using the AOS REST API.
 requirements:
-  - "aos-pyez >= 0.6.0"
+  - "aos-pyez >= 0.6.1"
 options:
   session:
     description:
@@ -43,77 +42,77 @@ options:
     required: true
   name:
     description:
-      - Name of the Rack Type to manage.
+      - Name of the HCL entry to manage.
         Only one of I(name), I(id) or I(content) can be set.
   id:
     description:
-      - AOS Id of the Rack Type to manage (can't be used to create a new Rack Type),
+      - AOS Id of the HCL entry to manage (can't be used to create a new item),
         Only one of I(name), I(id) or I(content) can be set.
   content:
     description:
-      - Datastructure of the Rack Type to create. The data can be in YAML / JSON or
+      - Datastructure of the HCL to create. The data can be in YAML / JSON or
         directly a variable. It's the same datastructure that is returned
         on success in I(value).
   state:
     description:
-      - Indicate what is the expected state of the Rack Type (present or not).
+      - Indicate what is the expected state of the item (present or not).
     default: present
     choices: ['present', 'absent']
 '''
 
 EXAMPLES = '''
 
-- name: "Delete a Rack Type by name"
-  aos_rack_type:
+- name: "Delete a HCL by name"
+  aos_hcl:
     session: "{{ aos_session }}"
-    name: "my-rack-type"
+    name: "my-hcl"
     state: absent
 
-- name: "Delete a Rack Type by id"
-  aos_rack_type:
+- name: "Delete a HCL by id"
+  aos_hcl:
     session: "{{ aos_session }}"
     id: "45ab26fc-c2ed-4307-b330-0870488fa13e"
     state: absent
 
-# Save a Rack Type to a file
+# Save a HCL to a file
 
-- name: "Access Rack Type 1/3"
-  aos_rack_type:
+- name: "Access HCL 1/3"
+  aos_hcl:
     session: "{{ aos_session }}"
-    name: "my-rack-type"
+    name: "my-hcl"
     state: present
-  register: rack_type
-- name: "Save Rack Type into a JSON file 2/3"
+  register: hcl
+- name: "Save HCL into a JSON file 2/3"
   copy:
-    content: "{{ rack_type.value | to_nice_json }}"
-    dest: rack_type_saved.json
-- name: "Save Rack Type into a YAML file 3/3"
+    content: "{{ hcl.value | to_nice_json }}"
+    dest: hcl_saved.json
+- name: "Save HCL into a YAML file 3/3"
   copy:
-    content: "{{ rack_type.value | to_nice_yaml }}"
-    dest: rack_type_saved.yaml
+    content: "{{ hcl.value | to_nice_yaml }}"
+    dest: hcl_saved.yaml
 
-- name: "Load Rack Type from a JSON file"
-  aos_rack_type:
+- name: "Load HCL from a JSON file"
+  aos_hcl:
     session: "{{ aos_session }}"
-    content: "{{ lookup('file', 'resources/rack_type_saved.json') }}"
+    content: "{{ lookup('file', 'resources/hcl_saved.json') }}"
     state: present
 
-- name: "Load Rack Type from a YAML file"
-  aos_rack_type:
+- name: "Load HCL from a YAML file"
+  aos_hcl:
     session: "{{ aos_session }}"
-    content: "{{ lookup('file', 'resources/rack_type_saved.yaml') }}"
+    content: "{{ lookup('file', 'resources/hcl_saved.yaml') }}"
     state: present
 '''
 
 RETURNS = '''
 name:
-  description: Name of the Rack Type
+  description: Name of the HCL
   returned: always
   type: str
   sample: AOS-1x25-1
 
 id:
-  description: AOS unique ID assigned to the Rack Type
+  description: AOS unique ID assigned to the HCL
   returned: always
   type: str
   sample: fcc4ac1c-e249-4fe7-b458-2138bfb44c06
@@ -125,75 +124,83 @@ value:
   sample: {'...'}
 '''
 
-import json
+import time
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.aos import get_aos_session, find_collection_item, do_load_resource, check_aos_version, content_to_dict
+
+from ansible.module_utils.aos import (
+    get_aos_session, find_collection_item,
+    do_load_resource, content_to_dict)
+
 
 #########################################################
 # State Processing
 #########################################################
-def rack_type_absent(module, aos, my_rack_type):
+
+def item_absent(module, item):
 
     margs = module.params
 
     # If the module do not exist, return directly
-    if my_rack_type.exists is False:
+    if item.exists is False:
         module.exit_json(changed=False,
                          name=margs['name'],
                          id=margs['id'],
                          value={})
 
-    # If not in check mode, delete Rack Type
+    # If not in check mode, delete HCL
     if not module.check_mode:
         try:
-            my_rack_type.delete()
+            # Need to way 1sec before a delete to workaround a current limitation in AOS
+            time.sleep(1)
+            item.delete()
         except:
-            module.fail_json(msg="An error occurred, while trying to delete the Rack Type")
+            module.fail_json(msg="An error occured, while trying to delete the HCL")
 
-    module.exit_json( changed=True,
-                      name=my_rack_type.name,
-                      id=my_rack_type.id,
-                      value={} )
+    module.exit_json(changed=True, name=item.name,
+                     id=item.id, value={})
 
-def rack_type_present(module, aos, my_rack_type):
+
+def item_present(module, item):
 
     margs = module.params
 
     if margs['content'] is not None:
 
         if 'display_name' in module.params['content'].keys():
-            do_load_resource(module, aos.RackTypes, module.params['content']['display_name'])
+            do_load_resource(module, item.collection, module.params['content']['display_name'])
         else:
             module.fail_json(msg="Unable to find display_name in 'content', Mandatory")
 
-    # if rack_type doesn't exist already, create a new one
-    if my_rack_type.exists is False and 'content' not in margs.keys():
+    # if item doesn't exist already, create a new one
+
+    if item.exists is False and 'content' not in margs.keys():
         module.fail_json(msg="'content' is mandatory for module that don't exist currently")
 
-    module.exit_json( changed=False,
-                      name=my_rack_type.name,
-                      id=my_rack_type.id,
-                      value=my_rack_type.value )
+    module.exit_json(changed=False, name=item.name,
+                     id=item.id, value=item.value)
+
 
 #########################################################
 # Main Function
 #########################################################
-def rack_type(module):
+
+def aos_hardware(module):
 
     margs = module.params
 
     try:
         aos = get_aos_session(module, margs['session'])
-    except:
-        module.fail_json(msg="Unable to login to the AOS server")
+    except Exception as exc:
+        module.fail_json(
+            msg="Unable to login to the AOS server: %s" % str(exc))
 
     item_name = False
     item_id = False
 
     if margs['content'] is not None:
 
-        content = content_to_dict(module, margs['content'] )
+        content = content_to_dict(module, margs['content'])
 
         if 'display_name' in content.keys():
             item_name = content['display_name']
@@ -206,44 +213,41 @@ def rack_type(module):
     elif margs['id'] is not None:
         item_id = margs['id']
 
-    #----------------------------------------------------
+    # ----------------------------------------------------
     # Find Object if available based on ID or Name
-    #----------------------------------------------------
-    my_rack_type = find_collection_item(aos.RackTypes,
-                                          item_name=item_name,
-                                          item_id=item_id)
+    # ----------------------------------------------------
 
-    #----------------------------------------------------
+    item = find_collection_item(aos.Hardware,
+                                item_name=item_name,
+                                item_id=item_id)
+
+    # ----------------------------------------------------
     # Proceed based on State value
-    #----------------------------------------------------
+    # ----------------------------------------------------
+
     if margs['state'] == 'absent':
-
-        rack_type_absent(module, aos, my_rack_type)
-
+        item_absent(module, item)
     elif margs['state'] == 'present':
+        item_present(module, item)
 
-        rack_type_present(module, aos, my_rack_type)
 
 def main():
     module = AnsibleModule(
         argument_spec=dict(
             session=dict(required=True, type="dict"),
-            name=dict(required=False ),
-            id=dict(required=False ),
+            name=dict(required=False),
+            id=dict(required=False),
             content=dict(required=False, type="json"),
-            state=dict( required=False,
-                        choices=['present', 'absent'],
-                        default="present")
+            state=dict(required=False,
+                       choices=['present', 'absent'],
+                       default="present")
         ),
-        mutually_exclusive = [('name', 'id', 'content')],
+        mutually_exclusive=[('name', 'id', 'content')],
         required_one_of=[('name', 'id', 'content')],
         supports_check_mode=True
     )
 
-    # Check if aos-pyez is present and match the minimum version
-    check_aos_version(module, '0.6.0')
-
-    rack_type(module)
+    aos_hardware(module)
 
 if __name__ == "__main__":
     main()
